Move VecCString to a separate module
